+++ pytdx/parser/get_finance_info.py
@@ -80,13 +80,10 @@ class GetFinanceInfo(BaseParser):
             weifenlirun,
             baoliu1,
             baoliu2
-        ) = struct.unpack("<IHHIIIIIIIIIIIIIIIIIIIIIIIIIIIIIIII", body_buf[pos:])
+        ) = struct.unpack("<fHHIIffffffffffffffffffffffffffffff", body_buf[pos:])
 
         def _get_v(v):
-            if v == 0:
-                return 0
-            else:
-                return get_volume(v)
+            return v
 
         return OrderedDict(
             [
@@ -129,3 +126,10 @@ class GetFinanceInfo(BaseParser):
                 ("baoliu2", _get_v(baoliu2))
             ]
         )
+
+if __name__ == '__main__':
+    import pprint
+    from pytdx.hq import TdxHq_API
+    api = TdxHq_API()
+    with api.connect():
+        pprint.pprint(api.get_finance_info(0, "000166"))